Refactor(member): Member 조회 단건/다건
다건같은 경우 현승님 코드를 활용했습니다.

+++ src/main/java/com/example/newsfeed/follow/repository/FollowRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 public interface FollowRepository extends JpaRepository<Follow, Long> {
 
-    long countByFollowingMemberId(Long id);
+    int countByFollowingMemberId(Long id);
 
     @Query("SELECT f FROM Follow f WHERE f.followerMember.id = :followerId AND f.followingMember.id = :followingId")
     Optional<Follow> findByFollowerIdAndFollowingId(
